added ranlib to Makefile, and changed print_len var from size_t to int to support negative values

+++ ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: gkhodizo <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2020/06/27 14:17:59 by gkhodizo          #+#    #+#             */
-/*   Updated: 2020/07/30 00:04:47 by gkhodizo         ###   ########.fr       */
+/*   Updated: 2020/07/30 17:16:01 by gkhodizo         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -35,7 +35,7 @@ typedef struct	s_fmt
 
 typedef struct	s_len
 {
-	size_t		print_len;
+	int			print_len;
 }				t_len;
 
 /*
